Adding implementation of TensoRF.
This CL implements only the 4D color (appearance) factorization. The code is adapted from the released pytorch code.

Given a colored point cloud, bake_colors optimizes a TensoRF approximation of the colors. This function is intended to be used to bake vertex colors for a surface extracted using the functions in surface.py.

PiperOrigin-RevId: 494860199